REFACTOR: Move models and params inside indexes.
