+++ services/indexes/avm/db_test.go
@@ -16,8 +16,7 @@ import (
 	"github.com/gocraft/dbr/v2"
 
 	"github.com/ava-labs/ortelius/cfg"
-	"github.com/ava-labs/ortelius/services"
-	"github.com/ava-labs/ortelius/services/models"
+	"github.com/ava-labs/ortelius/services/indexes/models"
 )
 
 type message struct {
