+++ services/indexes/avm/models.go
@@ -6,7 +6,7 @@ package avm
 import (
 	"time"
 
-	"github.com/ava-labs/ortelius/services/models"
+	"github.com/ava-labs/ortelius/services/indexes/models"
 )
 
 var (
